+++ nall/range.hpp
@@ -4,25 +4,25 @@ namespace nall {
 
 struct range_t {
   struct iterator {
-    iterator(int position, int step = 0) : position(position), step(step) {}
-    auto operator*() const -> int { return position; }
+    iterator(int64_t position, int64_t step = 0) : position(position), step(step) {}
+    auto operator*() const -> int64_t { return position; }
     auto operator!=(const iterator& source) const -> bool { return step > 0 ? position < source.position : position > source.position; }
     auto operator++() -> iterator& { position += step; return *this; }
 
   private:
-    int position;
-    const int step;
+    int64_t position;
+    const int64_t step;
   };
 
   struct reverse_iterator {
-    reverse_iterator(int position, int step = 0) : position(position), step(step) {}
-    auto operator*() const -> int { return position; }
+    reverse_iterator(int64_t position, int64_t step = 0) : position(position), step(step) {}
+    auto operator*() const -> int64_t { return position; }
     auto operator!=(const reverse_iterator& source) const -> bool { return step > 0 ? position > source.position : position < source.position; }
     auto operator++() -> reverse_iterator& { position -= step; return *this; }
 
   private:
-    int position;
-    const int step;
+    int64_t position;
+    const int64_t step;
   };
 
   auto begin() const -> iterator { return {origin, stride}; }
@@ -31,20 +31,20 @@ struct range_t {
   auto rbegin() const -> reverse_iterator { return {target - stride, stride}; }
   auto rend() const -> reverse_iterator { return {origin - stride}; }
 
-  int origin;
-  int target;
-  int stride;
+  int64_t origin;
+  int64_t target;
+  int64_t stride;
 };
 
-inline auto range(int size) {
+inline auto range(int64_t size) {
   return range_t{0, size, 1};
 }
 
-inline auto range(int offset, int size) {
+inline auto range(int64_t offset, int64_t size) {
   return range_t{offset, size, 1};
 }
 
-inline auto range(int offset, int size, int step) {
+inline auto range(int64_t offset, int64_t size, int64_t step) {
   return range_t{offset, size, step};
 }
