+++ nall/vector/compare.hpp
@@ -5,7 +5,7 @@ namespace nall {
 template<typename T> auto vector<T>::operator==(const vector<T>& source) const -> bool {
   if(this == &source) return true;
   if(size() != source.size()) return false;
-  for(uint n = 0; n < size(); n++) {
+  for(uint64_t n = 0; n < size(); n++) {
     if(operator[](n) != source[n]) return false;
   }
   return true;
