+++ nall/vector/modify.hpp
@@ -19,7 +19,7 @@ template<typename T> auto vector<T>::prepend(T&& value) -> void {
 template<typename T> auto vector<T>::prepend(const vector<T>& values) -> void {
   reserveLeft(size() + values.size());
   _pool -= values.size();
-  for(uint n : range(values)) new(_pool + n) T(values[n]);
+  for(uint64_t n : range(values)) new(_pool + n) T(values[n]);
   _left -= values.size();
   _size += values.size();
 }
@@ -27,7 +27,7 @@ template<typename T> auto vector<T>::prepend(const vector<T>& values) -> void {
 template<typename T> auto vector<T>::prepend(vector<T>&& values) -> void {
   reserveLeft(size() + values.size());
   _pool -= values.size();
-  for(uint n : range(values)) new(_pool + n) T(move(values[n]));
+  for(uint64_t n : range(values)) new(_pool + n) T(move(values[n]));
   _left -= values.size();
   _size += values.size();
 }
@@ -50,26 +50,26 @@ template<typename T> auto vector<T>::append(T&& value) -> void {
 
 template<typename T> auto vector<T>::append(const vector<T>& values) -> void {
   reserveRight(size() + values.size());
-  for(uint n : range(values.size())) new(_pool + _size + n) T(values[n]);
+  for(uint64_t n : range(values.size())) new(_pool + _size + n) T(values[n]);
   _right -= values.size();
   _size += values.size();
 }
 
 template<typename T> auto vector<T>::append(vector<T>&& values) -> void {
   reserveRight(size() + values.size());
-  for(uint n : range(values.size())) new(_pool + _size + n) T(move(values[n]));
+  for(uint64_t n : range(values.size())) new(_pool + _size + n) T(move(values[n]));
   _right -= values.size();
   _size += values.size();
 }
 
 //
 
-template<typename T> auto vector<T>::insert(uint offset, const T& value) -> void {
+template<typename T> auto vector<T>::insert(uint64_t offset, const T& value) -> void {
   if(offset == 0) return prepend(value);
   if(offset == size() - 1) return append(value);
   reserveRight(size() + 1);
   _size++;
-  for(int n = size() - 1; n > offset; n--) {
+  for(int64_t n = size() - 1; n > offset; n--) {
     _pool[n] = move(_pool[n - 1]);
   }
   new(_pool + offset) T(value);
@@ -77,21 +77,21 @@ template<typename T> auto vector<T>::insert(uint offset, const T& value) -> void
 
 //
 
-template<typename T> auto vector<T>::removeLeft(uint length) -> void {
+template<typename T> auto vector<T>::removeLeft(uint64_t length) -> void {
   if(length > size()) length = size();
   resizeLeft(size() - length);
 }
 
-template<typename T> auto vector<T>::removeRight(uint length) -> void {
+template<typename T> auto vector<T>::removeRight(uint64_t length) -> void {
   if(length > size()) length = size();
   resizeRight(size() - length);
 }
 
-template<typename T> auto vector<T>::remove(uint offset, uint length) -> void {
+template<typename T> auto vector<T>::remove(uint64_t offset, uint64_t length) -> void {
   if(offset == 0) return removeLeft(length);
   if(offset == size() - 1) return removeRight(length);
 
-  for(uint n = offset; n < size(); n++) {
+  for(uint64_t n = offset; n < size(); n++) {
     if(n + length < size()) {
       _pool[n] = move(_pool[n + length]);
     } else {
@@ -115,7 +115,7 @@ template<typename T> auto vector<T>::takeRight() -> T {
   return value;
 }
 
-template<typename T> auto vector<T>::take(uint offset) -> T {
+template<typename T> auto vector<T>::take(uint64_t offset) -> T {
   if(offset == 0) return takeLeft();
   if(offset == size() - 1) return takeRight();
