+++ nall/vector/utility.hpp
@@ -3,7 +3,7 @@
 namespace nall {
 
 template<typename T> auto vector<T>::fill(const T& value) -> void {
-  for(uint n : range(size())) _pool[n] = value;
+  for(uint64_t n : range(size())) _pool[n] = value;
 }
 
 template<typename T> auto vector<T>::sort(const function<bool (const T& lhs, const T& rhs)>& comparator) -> void {
@@ -12,24 +12,24 @@ template<typename T> auto vector<T>::sort(const function<bool (const T& lhs, con
 
 template<typename T> auto vector<T>::reverse() -> void {
   vector<T> reversed;
-  for(uint n : range(size())) reversed.prepend(_pool[n]);
+  for(uint64_t n : range(size())) reversed.prepend(_pool[n]);
   operator=(move(reversed));
 }
 
-template<typename T> auto vector<T>::find(const function<bool (const T& lhs)>& comparator) -> maybe<uint> {
-  for(uint n : range(size())) if(comparator(_pool[n])) return n;
+template<typename T> auto vector<T>::find(const function<bool (const T& lhs)>& comparator) -> maybe<uint64_t> {
+  for(uint64_t n : range(size())) if(comparator(_pool[n])) return n;
   return nothing;
 }
 
-template<typename T> auto vector<T>::find(const T& value) const -> maybe<uint> {
-  for(uint n : range(size())) if(_pool[n] == value) return n;
+template<typename T> auto vector<T>::find(const T& value) const -> maybe<uint64_t> {
+  for(uint64_t n : range(size())) if(_pool[n] == value) return n;
   return nothing;
 }
 
-template<typename T> auto vector<T>::findSorted(const T& value) const -> maybe<uint> {
-  int l = 0, r = size() - 1;
+template<typename T> auto vector<T>::findSorted(const T& value) const -> maybe<uint64_t> {
+  int64_t l = 0, r = size() - 1;
   while(l <= r) {
-    int m = l + (r - l >> 1);
+    int64_t m = l + (r - l >> 1);
     if(value == _pool[m]) return m;
     value < _pool[m] ? r = m - 1 : l = m + 1;
   }
@@ -37,11 +37,11 @@ template<typename T> auto vector<T>::findSorted(const T& value) const -> maybe<u
 }
 
 template<typename T> auto vector<T>::foreach(const function<void (const T&)>& callback) -> void {
-  for(uint n : range(size())) callback(_pool[n]);
+  for(uint64_t n : range(size())) callback(_pool[n]);
 }
 
 template<typename T> auto vector<T>::foreach(const function<void (uint, const T&)>& callback) -> void {
-  for(uint n : range(size())) callback(n, _pool[n]);
+  for(uint64_t n : range(size())) callback(n, _pool[n]);
 }
 
 }
